pyproject.toml
[project]
name = "smart-home-guard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "catboost>=1.2.10",
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "uvicorn[standard]>=0.43.0",
    "xgboost>=3.2.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
